Implement #102 - Incorrect Pixel-Depth Scaling
This commit fixes the incorrect scaling of depth values with the `near_far` parameter. Instead, the `near_far` paramater is only used for clipping camera geometry between a distance range. A new parameter `depth_distance` can now be used to specify the distance range for which camera geometry is mapped between lightness values 0-255.